+++ BidMachine/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## 3.0.1.0
+## 3.0.1.1
+* Requires minimum AppLovin MAX SDK version be 13.0.0.
+* Removed support for COPPA.
+* Remove deprecated native API usages.
 * Certified with BidMachine SDK 3.0.1.
 
 ## 3.0.0.0